v1.0

+++ calc_albsa.py
@@ -106,6 +106,11 @@ yrlist = []
 for i in range(yrbeg,yrend+1):
     yrlist.append(str(i))
 
+# for the NetCDF history
+def code_version():
+    cv = '1.0, 5/28/2025'
+    return cv
+
 # uses the Climate Data Store (CDS) API to download data from ECMWF
 def era5_downloader(yrlist,tname):
 
@@ -141,7 +146,7 @@ def era5_downloader(yrlist,tname):
             "18:00"
         ],
         "pressure_level": ["850"],
-        "grid": ["5/5"],
+        "grid": ["2.5/2.5"],
         "data_format": "netcdf",
         "download_format": "unarchived",
         "area": [85, 150, 40, -120]
@@ -155,14 +160,17 @@ def define_glob_atts():
 
     # key-value pairs are attributes names / attributes
     global_atts = {
-            "date_created"     :datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
             "title"            :"Aleutian Low-Beaufort Sea Anticyclone (ALBSA) index",
             "institution"      :"National Oceanic and Atmospheric Administration (NOAA) Physical Sciences Laboratory (PSL)",
+            "history"          :"created "+datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")+" with calc_albsa.py version "+code_version(),
             "file_creator"     :"Christopher J. Cox",
             "creator_email"    :"christopher.j.cox@noaa.gov", 
-            "Funding"          :"NOAA Physical Sciences Laboratory (PSL)",
-            "source"           :"ECMWF Reanalysis v5 (ERA5)", 
-            "methods"          :"Calculation based on the study of Cox et al. (2019).",
+            "funding"          :"NOAA Physical Sciences Laboratory (PSL)",
+            "source"           :"ECMWF Reanalysis v5 (ERA5)",
+            "license"          :"Creative Commons Attribution 4.0 License, CC 4.0",  
+            "methods"          :"Refer to Cox et al. (2019)",
+            "conventions"      :"CF-1.12",  
+            "commment"         :"",
             "keywords"         :"Arctic, climate index, snow, Alaska, atmospheric advection, sea ice",
             "acknowledgements" :"This product was generated using the Copernicus Climate Change Service information (CCCS/CDS, 2023) using data ECMWF Reanalysis v5 (ERA5) (Hersbach, et al. 2020; Hersbach et al., 2023).",
             "references"       :"""References:
@@ -196,9 +204,26 @@ def define_var_atts():
        
     # keys are place holders      {"variable","att_name","att_value"}
     var_atts = {
+
         "index_att1"            :("index", "long_name", "ALBSA index"),
         "index_att2"            :("index", "units", "meters"),
         "index_att3"            :("index", "standard_name", ""),
+
+        "scoord_att1"           :("s_coord", "long_name", "geopotential height at S coordinate"),
+        "scoord_att2"           :("s_coord", "units", "meters"),
+        "scoord_att3"           :("s_coord", "standard_name", "geopotential_height"),
+
+        "ncoord_att1"           :("n_coord", "long_name", "geopotential height at N coordinate"),
+        "ncoord_att2"           :("n_coord", "units", "meters"),
+        "ncoord_att3"           :("n_coord", "standard_name", "geopotential_height"),
+
+        "ecoord_att1"           :("e_coord", "long_name", "geopotential height at E coordinate"),
+        "ecoord_att2"           :("e_coord", "units", "meters"),
+        "ecoord_att3"           :("e_coord", "standard_name", "geopotential_height"),
+
+        "wcoord_att1"           :("w_coord", "long_name", "geopotential height at W coordinate"),
+        "wcoord_att2"           :("w_coord", "units", "meters"),
+        "wcoord_att3"           :("w_coord", "standard_name", "geopotential_height"),
     }
 
     return var_atts
@@ -209,8 +234,8 @@ def main():
     #   this needs to be batched. limits for CDS are 60K "items", which is 1 field x 1 var x 1 level x n time steps
     #   for this application at 4x/day, 20 years is about 30K. So lets do 20 year batches.
     batch_size = 5
-    #for i in range(0, len(yrlist), batch_size):
-    #    era5_downloader(yrlist[i:i+batch_size],"tmp_"+yrlist[i]+".nc")
+    for i in range(0, len(yrlist), batch_size):
+        era5_downloader(yrlist[i:i+batch_size],"tmp_"+yrlist[i]+".nc")
 
     # open the temporary ERA5 files using xarray and merge them into one xr.Dataset()
     file = xr.Dataset()
@@ -233,47 +258,60 @@ def main():
     laW = 55
     loW = 160
 
+    # convert the geopotential field [m^2/s^2] to geopotential height [m]
+    z_attrs = file['z'].attrs.copy() # save the attributes
+    grav = 9.80665 # gravity constant used by ECMWF
+    file['z'] = file['z']/grav
+    file['z'].attrs = z_attrs # reassign the attributes
+
     # 850 hPa daily mean GPH at 4 ALBSA coordinates.
     # The method is NN, but the data has been downloaded on pre-processed 2.5d grid, 
     #   which is a coarser, but divisible resampling of the native 0.25d grid.  
-    Scoord = file.sel(latitude=laS, longitude=loS, method="nearest")/9.81
-    Ncoord = file.sel(latitude=laN, longitude=loN, method="nearest")/9.81
-    Ecoord = file.sel(latitude=laE, longitude=loE, method="nearest")/9.81
-    Wcoord = file.sel(latitude=laW, longitude=loW, method="nearest")/9.81
+    Scoord = file.sel(latitude=laS, longitude=loS, method="nearest")
+    Ncoord = file.sel(latitude=laN, longitude=loN, method="nearest")
+    Ecoord = file.sel(latitude=laE, longitude=loE, method="nearest")
+    Wcoord = file.sel(latitude=laW, longitude=loW, method="nearest")
 
     # calculate ALBSA index
     #   Eq. (3) from Cox et al. (2019)
     albi = ( Ecoord.z - Wcoord.z ) - ( Ncoord.z - Scoord.z )
-    file = file.assign(index=albi.transpose())
     
+    # create vars for index variables
+    file = file.assign(index=albi.transpose())
+    file = file.assign(s_coord=Scoord.z.transpose())
+    file = file.assign(n_coord=Ncoord.z.transpose())
+    file = file.assign(e_coord=Ecoord.z.transpose())
+    file = file.assign(w_coord=Wcoord.z.transpose())
+
     # housekeeping
-    file["number"].drop_attrs()
+        # placeholder from CDS for ensembles. deceptive for reanalysis, in this case = 1
+    file["number"].drop_attrs() 
     file = file.drop_vars("number") 
+        # 4d to 3d bv squeezing the single-level pressure field
     file["z"] = file["z"].squeeze()
+        # use single 
     file["latitude"] = file["latitude"].astype("float32")
     file["longitude"] = file["longitude"].astype("float32")
 
     # if an existing data set is found, politely move it
-    if os.path.isfile(working_dir+fname):
-        shutil.move(working_dir+fname, working_dir+fname+".arch")
+    if os.path.isfile(working_dir+fname): shutil.move(working_dir+fname, working_dir+fname+".arch")
 
     # we will carry most of the attributes, but "index" is a new var, so we created some in define_var_atts()
     var_atts = define_var_atts()
-    for value in var_atts.values():
-        file[value[0]].attrs[value[1]] = value[2]
+    for value in var_atts.values(): file[value[0]].attrs = {} # clear current atts
+    for value in var_atts.values(): file[value[0]].attrs[value[1]] = value[2] # replace with desired atts
 
     # global attributes
     global_atts = define_glob_atts()
-    for key, value in global_atts.items():
-        file.attrs[key] = value
+    for key, value in global_atts.items(): file.attrs[key] = value
 
     # write back to new file
     file.to_netcdf(working_dir+fname)
     file.close()
 
     # delete the intermediary files
-    #for filename in glob.glob(working_dir+"tmp*.nc"):
-    #    os.remove(filename)
+    for filename in glob.glob(working_dir+"tmp*.nc"):
+        os.remove(filename)
 
 
 # executes main():
